Sort electrode states, "on" electrodes listed 1st
An electrode may appear twice in the list of modified electrode states
in cases where the same channel is mapped to multiple electrodes.

Sort electrode states with "on" electrodes listed first so the "on"
state will take precedence when the electrode controller plugin drops
duplicate states for the same electrode.