feat(zmq): use watch_plugin() to monitor socket
Prior to this commit, the ZeroMQ plugin socket was polled in the main GTK
thread.  However, this prevented responding to ZeroMQ plugin requests
while the GTK thread was blocked.

In this commit, use the `zmq_plugin.plugin.watch_plugin()` function to
monitor the ZeroMQ interface socket in a background thread.