fix(step-fields): make titles human-readable
Prior to this commit, the inferred column title in the protocol grid for
the `repeat_duration_s` step field was "Repeat duration s", which is
somewhat awkward.

In this commit, explicitly set the displayed title of the
`repeat_duration_s` step field to `"Repeat duration (s)"`, which is more
human-readable.  Note that the underlying step field _name_ remains the
same (i.e., `repeat_duration_s`), maintaining backwards compatibility with
existing protocols.